# Remove closure allocations from JintFunctionDefinition