## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|dromaeo-3d-cube|15.385 ms|0.2515 ms|6305.56 KB|
| **New** |	|	| **15.568 ms (+1%)** | **0.1567 ms** | **6215.83 KB (-1%)** |
| Old |Run|dromaeo-core-eval|3.041 ms|0.0381 ms|327.41 KB|
| **New** |	|	| **3.122 ms (+3%)** | **0.0254 ms** | **327.06 KB (0%)** |
| Old |Run|dromaeo-object-array|31.828 ms|0.2936 ms|96259.1 KB|
| **New** |	|	| **31.940 ms (0%)** | **0.1914 ms** | **96258.38 KB (0%)** |
| Old |Run|droma(...)egexp [21]|123.952 ms|2.4561 ms|149449.74 KB|
| **New** |	|	| **124.259 ms (0%)** | **2.4668 ms** | **151579.51 KB (+1%)** |
| Old |Run|droma(...)tring [21]|238.855 ms|5.2330 ms|1315958.49 KB|
| **New** |	|	| **244.141 ms (+2%)** | **6.9506 ms** | **1315860.24 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|35.305 ms|0.4351 ms|2367.42 KB|
| **New** |	|	| **34.810 ms (-1%)** | **0.5486 ms** | **2367.03 KB (0%)** |
| Old |Run|dromaeo-3d-cube|15.041 ms|0.1465 ms|5967.31 KB|
| **New** |	|	| **14.848 ms (-1%)** | **0.1598 ms** | **5877.58 KB (-2%)** |
| Old |Run|dromaeo-core-eval|3.066 ms|0.0264 ms|312.35 KB|
| **New** |	|	| **3.264 ms (+6%)** | **0.0227 ms** | **311.99 KB (0%)** |
| Old |Run|dromaeo-object-array|32.979 ms|0.4518 ms|96211.68 KB|
| **New** |	|	| **31.887 ms (-3%)** | **0.4474 ms** | **96210.99 KB (0%)** |
| Old |Run|droma(...)egexp [21]|96.108 ms|1.9185 ms|152222.83 KB|
| **New** |	|	| **88.918 ms (-7%)** | **1.7744 ms** | **150412.28 KB (-1%)** |
| Old |Run|droma(...)tring [21]|243.834 ms|4.8618 ms|1315744.62 KB|
| **New** |	|	| **242.248 ms (-1%)** | **6.3641 ms** | **1315700.62 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|34.412 ms|0.4386 ms|2268.26 KB|
| **New** |	|	| **34.364 ms (0%)** | **0.3516 ms** | **2267.87 KB (0%)** |


## Jint.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|122.85 ms|1.239 ms|45166.97 KB|
| **New** |	|	| **123.38 ms (0%)** | **1.403 ms** | **44653.93 KB (-1%)** |
| Old |Run|3d-morph|110.68 ms|0.736 ms|46258.12 KB|
| **New** |	|	| **111.42 ms (+1%)** | **1.204 ms** | **46257.53 KB (0%)** |
| Old |Run|3d-raytrace|117.33 ms|1.259 ms|86419.7 KB|
| **New** |	|	| **116.28 ms (-1%)** | **1.775 ms** | **84207.66 KB (-3%)** |
| Old |Run|access-binary-trees|66.71 ms|1.304 ms|62249.15 KB|
| **New** |	|	| **64.45 ms (-3%)** | **0.746 ms** | **57318.84 KB (-8%)** |
| Old |Run|access-fannkuch|269.99 ms|2.952 ms|103.82 KB|
| **New** |	|	| **272.92 ms (+1%)** | **3.480 ms** | **103.78 KB (0%)** |
| Old |Run|access-nbody|128.52 ms|1.207 ms|53295.29 KB|
| **New** |	|	| **128.38 ms (0%)** | **1.167 ms** | **53117.14 KB (0%)** |
| Old |Run|access-nsieve|90.79 ms|0.946 ms|17142.64 KB|
| **New** |	|	| **90.80 ms (0%)** | **0.727 ms** | **17142.49 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|98.41 ms|1.170 ms|61948.44 KB|
| **New** |	|	| **98.24 ms (0%)** | **1.180 ms** | **56948.4 KB (-8%)** |
| Old |Run|bitops-bits-in-byte|150.54 ms|1.540 ms|40544.67 KB|
| **New** |	|	| **148.96 ms (-1%)** | **1.652 ms** | **37044.63 KB (-9%)** |
| Old |Run|bitops-bitwise-and|84.62 ms|1.058 ms|55939.01 KB|
| **New** |	|	| **84.76 ms (0%)** | **0.764 ms** | **55939.01 KB (0%)** |
| Old |Run|bitops-nsieve-bits|145.31 ms|2.736 ms|53916.37 KB|
| **New** |	|	| **138.66 ms (-5%)** | **1.851 ms** | **53916.29 KB (0%)** |
| Old |Run|contr(...)rsive [21]|77.20 ms|1.515 ms|92772.92 KB|
| **New** |	|	| **77.20 ms (0%)** | **1.261 ms** | **83183.56 KB (-10%)** |
| Old |Run|crypto-aes|84.26 ms|1.576 ms|10729.22 KB|
| **New** |	|	| **84.41 ms (0%)** | **1.077 ms** | **10341.31 KB (-4%)** |
| Old |Run|crypto-md5|71.44 ms|0.758 ms|82176.41 KB|
| **New** |	|	| **69.40 ms (-3%)** | **1.195 ms** | **77797.5 KB (-5%)** |
| Old |Run|crypto-sha1|68.84 ms|0.691 ms|68861.17 KB|
| **New** |	|	| **67.90 ms (-1%)** | **1.013 ms** | **64485.15 KB (-6%)** |
| Old |Run|date-format-tofte|67.77 ms|0.598 ms|54161.32 KB|
| **New** |	|	| **66.54 ms (-2%)** | **0.851 ms** | **53477.73 KB (-1%)** |
| Old |Run|date-format-xparb|39.28 ms|0.749 ms|26400.21 KB|
| **New** |	|	| **38.11 ms (-3%)** | **0.744 ms** | **24992.57 KB (-5%)** |
| Old |Run|math-cordic|214.22 ms|3.227 ms|86859.41 KB|
| **New** |	|	| **223.02 ms (+4%)** | **2.089 ms** | **81976.09 KB (-6%)** |
| Old |Run|math-partial-sums|77.12 ms|1.001 ms|49367.9 KB|
| **New** |	|	| **77.27 ms (0%)** | **1.241 ms** | **49367.66 KB (0%)** |
| Old |Run|math-spectral-norm|86.40 ms|1.429 ms|56617.53 KB|
| **New** |	|	| **80.34 ms (-7%)** | **1.543 ms** | **51826.74 KB (-8%)** |
| Old |Run|regexp-dna|102.67 ms|1.791 ms|16826.82 KB|
| **New** |	|	| **96.15 ms (-6%)** | **1.742 ms** | **16828.59 KB (0%)** |
| Old |Run|string-base64|52.07 ms|0.632 ms|3156.22 KB|
| **New** |	|	| **51.68 ms (-1%)** | **0.698 ms** | **3156.06 KB (0%)** |
| Old |Run|string-fasta|114.49 ms|2.138 ms|104786.35 KB|
| **New** |	|	| **115.00 ms (0%)** | **2.109 ms** | **102598.66 KB (-2%)** |
| Old |Run|string-tagcloud|53.54 ms|0.707 ms|42219.39 KB|
| **New** |	|	| **51.34 ms (-4%)** | **0.447 ms** | **40456.12 KB (-4%)** |
| Old |Run|string-unpack-code|51.93 ms|0.766 ms|73810.43 KB|
| **New** |	|	| **51.31 ms (-1%)** | **0.885 ms** | **69834.8 KB (-5%)** |
| Old |Run|strin(...)input [21]|43.31 ms|0.840 ms|20687.13 KB|
| **New** |	|	| **42.54 ms (-2%)** | **0.537 ms** | **19905.84 KB (-4%)** |